requirements.txt
tensorflow==2.12.0  # Cambia según la versión que estés usando
numpy>=1.23.0
opencv-python>=4.7.0
Pillow>=9.5.0
streamlit>=1.25.0

